types: hide Span u32 fields behind usize accessors
Span stores offsets as u32 to keep the struct at 16 bytes (every AST
leaf carries one, and Nix sources are far below 4 GiB). The storage is
deliberate, but exposing raw u32 fields forced 'span.start as usize'
and 'span.start..span.end' at every call site.

Make the fields private and add const accessors returning usize:
start, end, start_line, end_line, len, is_empty, range. Update all
callers; ErrorContext::position now takes usize so span.start() is
passed directly.

+++ src/parser/path_uri.rs
@@ -107,7 +107,7 @@ impl Parser {
                 && text.ends_with('/')
             {
                 // Point to the trailing slash, not the start of the path
-                let current_pos = p.lexer.current_pos().start as usize;
+                let current_pos = p.lexer.current_pos().start();
                 let slash_pos = Span::new(current_pos.saturating_sub(1), current_pos);
                 return Err(ParseError::invalid(
                     slash_pos,
